Move sandbox files
