+++ .eslintrc.json
@@ -562,19 +562,19 @@
                 "ext/js/core/api-map.js",
                 "ext/js/core/extension-error.js",
                 "ext/js/core/json.js",
-                "ext/js/data/sandbox/anki-note-data-creator.js",
+                "ext/js/data/anki-note-data-creator.js",
                 "ext/js/dictionary/dictionary-data-util.js",
-                "ext/js/display/sandbox/pronunciation-generator.js",
-                "ext/js/display/sandbox/structured-content-generator.js",
-                "ext/js/dom/sandbox/css-style-applier.js",
+                "ext/js/display/pronunciation-generator.js",
+                "ext/js/display/structured-content-generator.js",
+                "ext/js/dom/css-style-applier.js",
                 "ext/js/language/ja/japanese.js",
                 "ext/js/language/text-utilities.js",
-                "ext/js/templates/sandbox/anki-template-renderer-content-manager.js",
-                "ext/js/templates/sandbox/anki-template-renderer.js",
-                "ext/js/templates/sandbox/template-renderer-frame-api.js",
-                "ext/js/templates/sandbox/template-renderer-frame-main.js",
-                "ext/js/templates/sandbox/template-renderer-media-provider.js",
-                "ext/js/templates/sandbox/template-renderer.js"
+                "ext/js/templates/anki-template-renderer-content-manager.js",
+                "ext/js/templates/anki-template-renderer.js",
+                "ext/js/templates/template-renderer-frame-api.js",
+                "ext/js/templates/template-renderer-frame-main.js",
+                "ext/js/templates/template-renderer-media-provider.js",
+                "ext/js/templates/template-renderer.js"
             ],
             "env": {
                 "webextensions": false
@@ -630,7 +630,7 @@
                 "ext/js/data/json-schema.js",
                 "ext/js/data/options-util.js",
                 "ext/js/data/permissions-util.js",
-                "ext/js/data/sandbox/array-buffer-util.js",
+                "ext/js/data/array-buffer-util.js",
                 "ext/js/dictionary/dictionary-database.js",
                 "ext/js/dom/native-simple-dom-parser.js",
                 "ext/js/dom/simple-dom-parser.js",
