+++ dev/jsconfig.json
@@ -63,7 +63,7 @@
         "../ext/js/data/database.js",
         "../ext/js/data/json-schema.js",
         "../ext/js/general/cache-map.js",
-        "../ext/js/data/sandbox/anki-note-data-creator.js",
+        "../ext/js/data/anki-note-data-creator.js",
         "../ext/js/general/cache-map.js",
         "../ext/js/general/regex-util.js",
         "../ext/js/general/text-source-map.js",
@@ -71,7 +71,7 @@
         "../ext/js/dictionary/dictionary-importer.js",
         "../ext/js/dictionary/dictionary-database.js",
         "../ext/js/dictionary/dictionary-data-util.js",
-        "../ext/js/language/sandbox/japanese-util.js",
+        "../ext/js/language/japanese-util.js",
         "../ext/js/language/translator.js",
         "../ext/js/media/media-util.js",
         "../types/dev/**/*.ts",
