+++ docs/templates.md
@@ -3,7 +3,7 @@
 ## Helpers
 
 Yomitan supports several custom Handlebars helpers for rendering templates.
-The source code for these templates can be found [here](../ext/js/templates/sandbox/anki-template-renderer.js).
+The source code for these templates can be found [here](../ext/js/templates/anki-template-renderer.js).
 
 ### `dumpObject`
