+++ ext/js/data/string-util.js
