+++ ext/js/templates/anki-template-renderer-content-manager.js
