+++ ext/js/templates/template-renderer-frame-main.js
