+++ ext/js/templates/template-renderer-media-provider.js
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import {Handlebars} from '../../../lib/handlebars.js';
+import {Handlebars} from '../../lib/handlebars.js';
 
 export class TemplateRendererMediaProvider {
     constructor() {
