+++ ext/template-renderer.html
@@ -11,7 +11,7 @@
     <link rel="icon" type="image/png" href="/images/icon48.png" sizes="48x48">
     <link rel="icon" type="image/png" href="/images/icon64.png" sizes="64x64">
     <link rel="icon" type="image/png" href="/images/icon128.png" sizes="128x128">
-    <script src="/js/templates/sandbox/template-renderer-frame-main.js" type="module"></script>
+    <script src="/js/templates/template-renderer-frame-main.js" type="module"></script>
 </head>
 <body>
